requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
asyncio>=3.4.3
Pillow>=10.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pytz>=2023.3
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
lxml>=4.9.3
matplotlib>=3.8.0
numpy>=1.25.0
pandas>=2.1.0
flask>=3.0.0
flask-cors>=4.0.0
psutil>=5.9.0
